# src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: CrossUniversityLibrary
  description: OpenAPI documentation for RESTful Service CrossUniversityLibrary
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Localhost Server URL
  - url: http://25.59.220.248:8080
    description: Hamachi Server URL
security:
  - bearerAuth: []
paths:
  /user/update-profile:
    put:
      tags:
        - User Controller
      summary: Update the user profile by changing the username
      description: >-
        This API endpoint is accessible to authenticated users, allowing them to
        update their user profile by changing the username.
      operationId: updateProfile
      parameters:
        - name: userName
          in: query
          description: New username to be set
          required: true
          schema:
            type: string
      responses:
        '200':
          description: >-
            The request was successful. The server has processed the request and
            the operation was completed without errors.
          content:
            application/json:
              example:
                id: 1
                userName: User Name
                email: email@example.com
                space: 1024
                university:
                  id: 1
                  title: university_title
                  domain: university_domain
                ownLibraries:
                  - id: 3
                    title: library1_title
                    topic: library1_topic
                    libraryAccess: false
                subscribedLibraries:
                  - id: 2
                    title: library2_title
                    topic: library2_topic
                    libraryAccess: true
        '401':
          description: >-
            Access to this resource requires authentication. Please log in to
            your account to gain the necessary permissions and access the
            requested resource.
          content:
            application/json:
              example:
                message: Unauthorized example message
                httpStatus: UNAUTHORIZED
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
        '403':
          description: >-
            Access to the requested resource is forbidden. The authenticated
            user lacks the necessary permissions to perform the requested
            operation.
          content:
            application/json:
              example:
                message: Forbidden example message
                httpStatus: FORBIDDEN
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
...
  /user/confirm-old-password:
    post:
      tags:
        - User Controller
      summary: >-
        Confirm the old password before initiating the password-changing
        procedure
      description: >-
        This API endpoint is accessible to authenticated users, allowing them to
        confirm their old password before initiating the password-changing
        procedure. The user needs to provide their current password for
        confirmation. After successful confirmation, a secret code is sent to
        the user's email for further identification in the password-changing
        process.
      operationId: confirmOldPassword
      parameters:
        - name: confirmedPassword
          in: query
          description: Current password to be confirmed
          required: true
          schema:
            type: string
      responses:
        '204':
          description: >-
            The request was successful, but there is no additional content to
            return. The server has processed the request, and no response body
            is provided.
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: >-
            Access to this resource requires authentication. Please log in to
            your account to gain the necessary permissions and access the
            requested resource.
          content:
            application/json:
              example:
                message: Unauthorized example message
                httpStatus: UNAUTHORIZED
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
        '403':
          description: >-
            Access to the requested resource is forbidden. The authenticated
            user lacks the necessary permissions to perform the requested
            operation.
          content:
            application/json:
              example:
                message: Forbidden example message
                httpStatus: FORBIDDEN
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
  /user/change-password:
    post:
      tags:
        - User Controller
      summary: Change the password with confirming the identity using secret code
      description: >-
        This API endpoint is accessible to authenticated users, allowing them to
        change their password after confirming their identity with a copied
        secret code. Once the user has received the secret code via email and
        copied it, they provide the code along with the new password to this
        method for identity confirmation. After successful confirmation, the
        user can change their old password to the new one.
      operationId: changePassword
      parameters:
        - name: copiedSecretCode
          in: query
          description: Secret code copied from the email
          required: true
          schema:
            type: string
        - name: newPassword
          in: query
          description: New password to be set
          required: true
          schema:
            type: string
      responses:
        '204':
          description: >-
            The request was successful, but there is no additional content to
            return. The server has processed the request, and no response body
            is provided.
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: >-
            Access to this resource requires authentication. Please log in to
            your account to gain the necessary permissions and access the
            requested resource.
          content:
            application/json:
              example:
                message: Unauthorized example message
                httpStatus: UNAUTHORIZED
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
        '403':
          description: >-
            Access to the requested resource is forbidden. The authenticated
            user lacks the necessary permissions to perform the requested
            operation.
          content:
            application/json:
              example:
                message: Forbidden example message
                httpStatus: FORBIDDEN
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
  /university/register-university:
    post:
      tags:
        - University Controller
      summary: University registration by global administrators
      description: >-
        This API endpoint is accessible to users with the role of global
        administrators, allowing them to register new universities in the
        system. To register a university, users need to provide the title and
        domain information through a UniversityRegisterRequest in the request
        body.
      operationId: registerUniversity
      parameters:
        - name: title
          in: query
          required: true
          schema:
            type: string
        - name: domain
          in: query
          required: true
          schema:
            type: string
      responses:
        '201':
          description: >-
            The resource has been successfully created. The server has processed
            the request to create a new resource, and the operation was
            completed without errors.
          content:
            application/json:
              example:
                id: 1
                title: title_example
                domain: domain.example
        '400':
          description: >-
            The server could not understand the request due to invalid or
            malformed parameters.
          content:
            application/json:
              example:
                message: Bad request example message
                httpStatus: BAD_REQUEST
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
...
  /user/{email}/profile:
    get:
      tags:
        - User Controller
      summary: Retrieve the profile information of a user by email
      description: >-
        This API endpoint is accessible to authenticated users, allowing them to
        retrieve and view the profile information of another user based on their
        email. The profile information includes the user's ID, username, email,
        available space, associated university details, and lists of owned and
        subscribed libraries.
      operationId: getUserProfileByEmail
      parameters:
        - name: email
          in: path
          description: >-
            Email address of the user for whom the profile information is
            requested
          required: true
          schema:
            type: string
      responses:
        '200':
          description: >-
            The request was successful. The server has processed the request and
            the operation was completed without errors.
          content:
            application/json:
              example:
                id: 1
                userName: User Name
                email: email@example.com
                space: 1024
                university:
                  id: 1
                  title: university_title
                  domain: university_domain
                ownLibraries:
                  - id: 3
                    title: library1_title
                    topic: library1_topic
                    libraryAccess: false
                subscribedLibraries:
                  - id: 2
                    title: library2_title
                    topic: library2_topic
                    libraryAccess: true
        '401':
          description: >-
            Access to this resource requires authentication. Please log in to
            your account to gain the necessary permissions and access the
            requested resource.
          content:
            application/json:
              example:
                message: Unauthorized example message
                httpStatus: UNAUTHORIZED
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
        '403':
          description: >-
            Access to the requested resource is forbidden. The authenticated
            user lacks the necessary permissions to perform the requested
            operation.
          content:
            application/json:
              example:
                message: Forbidden example message
                httpStatus: FORBIDDEN
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
        '404':
          description: >-
            The requested resource could not be found on the server. No data
            exists for this specific request. Please verify the URL and ensure
            that the resource exists.
          content:
            application/json:
              example:
                message: Not found example message
                httpStatus: NOT_FOUND
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
  /user/profile:
    get:
      tags:
        - User Controller
      summary: Retrieve the profile information of the authenticated user
      description: >-
        This API endpoint is accessible to authenticated users, allowing them to
        retrieve and view their own profile information. The profile information
        includes the user's ID, username, email, available space, associated
        university details, and lists of owned and subscribed libraries.
      operationId: profile
      responses:
        '200':
          description: >-
            The request was successful. The server has processed the request and
            the operation was completed without errors.
          content:
            application/json:
              example:
                id: 1
                userName: User Name
                email: email@example.com
                space: 1024
                university:
                  id: 1
                  title: university_title
                  domain: university_domain
                ownLibraries:
                  - id: 3
                    title: library1_title
                    topic: library1_topic
                    libraryAccess: false
                subscribedLibraries:
                  - id: 2
                    title: library2_title
                    topic: library2_topic
                    libraryAccess: true
        '401':
          description: >-
            Access to this resource requires authentication. Please log in to
            your account to gain the necessary permissions and access the
            requested resource.
          content:
            application/json:
              example:
                message: Unauthorized example message
                httpStatus: UNAUTHORIZED
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
        '403':
          description: >-
            Access to the requested resource is forbidden. The authenticated
            user lacks the necessary permissions to perform the requested
            operation.
          content:
            application/json:
              example:
                message: Forbidden example message
                httpStatus: FORBIDDEN
                timestamp: '2024-01-04T11:57:16.835Z'
                path: http://exception/example/path
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT authentication
      scheme: bearer
      bearerFormat: JWT